+++ Source/ACE.DatLoader/FileTypes/Animation.cs
@@ -10,18 +10,17 @@ namespace ACE.DatLoader.FileTypes
     /// Special thanks to Dan Skorupski for his work on Bael'Zharon's Respite, which helped fill in some of the gaps https://github.com/boardwalk/bzr
     /// </summary>
     [DatFileType(DatFileType.Animation)]
-    public class Animation : IUnpackable
+    public class Animation : FileType
     {
-        public uint Id { get; private set; }
         public uint Bitfield { get; private set; }
         public uint NumParts { get; private set; }
         public uint NumFrames { get; private set; }
         public List<Frame> PosFrames { get; } = new List<Frame>();
         public List<AnimationFrame> PartFrames { get; } = new List<AnimationFrame>();
 
-        public void Unpack(BinaryReader reader)
+        public override void Unpack(BinaryReader reader)
         {
-            Id = reader.ReadUInt32();
+            Id          = reader.ReadUInt32();
             Bitfield    = reader.ReadUInt32();
             NumParts    = reader.ReadUInt32();
             NumFrames   = reader.ReadUInt32();
@@ -35,26 +34,6 @@ namespace ACE.DatLoader.FileTypes
                 animationFrame.Unpack(reader, NumParts);
                 PartFrames.Add(animationFrame);
             }
-        }
-
-        public static Animation ReadFromDat(uint fileId)
-        {
-            // Check the FileCache so we don't need to hit the FileSystem repeatedly
-            if (DatManager.PortalDat.FileCache.TryGetValue(fileId, out var result))
-                return (Animation)result;
-
-            DatReader datReader = DatManager.PortalDat.GetReaderForFile(fileId);
-
-            var obj = new Animation();
-
-            using (var memoryStream = new MemoryStream(datReader.Buffer))
-            using (var reader = new BinaryReader(memoryStream))
-                obj.Unpack(reader);
-
-            // Store this object in the FileCache
-            DatManager.PortalDat.FileCache[fileId] = obj;
-
-            return obj;
         }
     }
 }