+++ Source/ACE.DatLoader/FileTypes/SpellTable.cs
@@ -7,41 +7,20 @@ using ACE.DatLoader.Entity;
 namespace ACE.DatLoader.FileTypes
 {
     [DatFileType(DatFileType.SpellTable)]
-    public class SpellTable : IUnpackable
+    public class SpellTable : FileType
     {
-        private const uint FILE_ID = 0x0E00000E;
+        internal const uint FILE_ID = 0x0E00000E;
 
         public Dictionary<uint, SpellBase> Spells { get; } = new Dictionary<uint, SpellBase>();
 
-        public void Unpack(BinaryReader reader)
+        public override void Unpack(BinaryReader reader)
         {
-            reader.BaseStream.Position += 4; // Skip the ID. We know what it is.
+            Id = reader.ReadUInt32();
 
             Spells.UnpackPackedHashTable(reader);
 
             // TODO: I don't know why, but the reader doesn't align properly with the length here
             reader.BaseStream.Position = reader.BaseStream.Length;
-        }
-
-        public static SpellTable ReadFromDat()
-        {
-            // Check the FileCache so we don't need to hit the FileSystem repeatedly
-            if (DatManager.PortalDat.FileCache.TryGetValue(FILE_ID, out var result))
-                return (SpellTable)result;
-
-            // Create the datReader for the proper file
-            DatReader datReader = DatManager.PortalDat.GetReaderForFile(FILE_ID);
-
-            var obj = new SpellTable();
-
-            using (var memoryStream = new MemoryStream(datReader.Buffer))
-            using (var reader = new BinaryReader(memoryStream))
-                obj.Unpack(reader);
-
-            // Store this object in the FileCache
-            DatManager.PortalDat.FileCache[FILE_ID] = obj;
-
-            return obj;
         }
 
         /// <summary>
@@ -72,13 +51,9 @@ namespace ACE.DatLoader.FileTypes
         /// <summary>
         /// Returns the correct spell formula, which is hashed from a player's account name
         /// </summary>
-        /// <param name="spellId"></param>
-        /// <param name="accountName"></param>
-        /// <returns>A list of spellcomp ids</returns>
-        public static List<uint> GetSpellFormula(uint spellId, string accountName)
+        public static List<uint> GetSpellFormula(SpellTable spellTable, uint spellId, string accountName)
         {
-            SpellTable s = ReadFromDat();
-            SpellBase spell = s.Spells[spellId];
+            SpellBase spell = spellTable.Spells[spellId];
            
             switch (spell.FormulaVersion)
             {